requirements.txt
flask
pillow
clarifai==2.0.20
python-dotenv
